refactor: 고정비 규칙 기간/상태 판단 로직을 유틸로 분리 #115
