+++ fe/src/components/fixed-costs/FixedCostItem.tsx
@@ -4,7 +4,8 @@ import { Item, ItemContent, ItemTitle } from '@/components/ui/item';
 import { cn } from '@/lib/utils';
 import { ChevronRight } from 'lucide-react';
 import { IFixedRule } from '@/types/fixed-costs';
-import { formatFixedRuleCycle, isEndedFixedRule } from '@/utils/fixed-costs';
+import { formatFixedRuleCycle } from '@/utils/fixed-costs/fixedCosts';
+import { isEndedFixedRule } from '@/utils/fixed-costs/rule';
 import { THEME_COLOR } from '@/constants/colors';
 import { CATEGORIES } from '@/constants/categories';
 import { Badge } from '../ui/badge';
